backend/chat_app/serializers.py
from .models import Message
from rest_framework import serializers


class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = ['user', 'receiver',  'content','read', 'timestamp']
