backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
dj-database-url==1.3.0
Django==4.1.7
django-cors-headers==4.3.1
django-livereload-server==0.5.1
django-livesync==0.5
django-redis==5.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
Pillow==9.5.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyotp==2.9.0
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
watchdog==3.0.0
websocket-client==1.7.0
whitenoise==6.4.0
zope.interface==6.1
psycopg2==2.9.3